+++ pom.xml
@@ -124,6 +124,7 @@
         <module>rest-testing</module>
         <module>resteasy</module>
         <module>rxjava</module>
+        <module>spring-swagger-codegen</module>
 
         <module>selenium-junit-testng</module>
         <module>solr</module>
