+++ spring-swagger-codegen/spring-swagger-codegen-app/pom.xml
@@ -2,8 +2,8 @@
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
-	<groupId>com.baeldung.swagger.codegen</groupId>
-	<artifactId>petstore-web</artifactId>
+	<groupId>com.baeldung</groupId>
+	<artifactId>spring-swagger-codegen-app</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 
 	<parent>
