+++ spring-swagger-codegen/spring-swagger-codegen-app/src/main/java/com/baeldung/petstore/app/DefaultPetService.java
@@ -1,4 +1,4 @@
-package com.baeldung.petstore.web;
+package com.baeldung.petstore.app;
 
 import java.util.Arrays;
 import java.util.List;
