+++ spring-swagger-codegen/spring-swagger-codegen-app/src/main/java/com/baeldung/petstore/app/PetController.java
@@ -1,4 +1,4 @@
-package com.baeldung.petstore.web;
+package com.baeldung.petstore.app;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -14,10 +14,10 @@ public class PetController {
 
     @RequestMapping("/pets")
     @ResponseBody
-    public String listAllPets() {
-        StringBuilder sb = new StringBuilder("<h1>Pending pets:</h1>");
+    public String listAvailablePets() {
+        StringBuilder sb = new StringBuilder("<h1>Available pets:</h1>");
         sb.append("<ul>");
-        petService.findPendingPets()
+        petService.findAvailablePets()
           .forEach( p -> sb.append("<li>" + p.getName() + "</li>"));
         sb.append("</ul>");
         return sb.toString();
