+++ spring-swagger-codegen/spring-swagger-codegen-app/src/main/java/com/baeldung/petstore/app/PetStoreApplication.java
@@ -1,4 +1,4 @@
-package com.baeldung.petstore.web;
+package com.baeldung.petstore.app;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
