+++ spring-swagger-codegen/spring-swagger-codegen-app/src/main/java/com/baeldung/petstore/app/PetStoreIntegrationConfig.java
@@ -1,11 +1,10 @@
-package com.baeldung.petstore.web;
+package com.baeldung.petstore.app;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 import com.baeldung.petstore.client.api.PetApi;
 import com.baeldung.petstore.client.invoker.ApiClient;
-import com.baeldung.petstore.client.invoker.auth.OAuth;
 
 @Configuration
 public class PetStoreIntegrationConfig {
@@ -18,10 +17,6 @@ public class PetStoreIntegrationConfig {
     @Bean
     public ApiClient apiClient() {
         ApiClient apiClient = new ApiClient();
-        
-        OAuth petStoreAuth = (OAuth) apiClient.getAuthentication("petstore_auth");
-        petStoreAuth.setAccessToken("special-key");
-
         return apiClient;
     }
