--- spring-swagger-codegen/spring-swagger-codegen-spring-app/src/main/java/com/baeldung/petstore/web/PetService.java
@@ -1,15 +1,0 @@
-package com.baeldung.petstore.web;
-
-import java.util.List;
-
-import com.baeldung.petstore.client.model.Pet;
-
-public interface PetService {
-
-    List<Pet> findAvailablePets();
-    
-    List<Pet> findPendingPets();
-    
-    List<Pet> findSoldPets();
-    
-}
